requirements.txt
streamlit
pandas
pdfplumber
pypdfium2
openpyxl
Pillow
pytesseract
python-dateutil
gspread
google-auth